+++ extension.php
@@ -44,8 +44,8 @@ class ArticleSummaryExtension extends Minz_Extension
       FreshRSS_Context::$user_conf->oai_key = Minz_Request::param('oai_key', '');
       FreshRSS_Context::$user_conf->oai_model = Minz_Request::param('oai_model', '');
       FreshRSS_Context::$user_conf->oai_prompt = Minz_Request::param('oai_prompt', '');
-      FreshRSS_Context::$user_conf->oai_max_tokens = Minz_Request::param('oai_max_tokens', 2048);
-      FreshRSS_Context::$user_conf->oai_temperature = Minz_Request::param('oai_temperature', 0.7);
+      FreshRSS_Context::$user_conf->oai_max_tokens = (int)Minz_Request::param('oai_max_tokens', 2048);
+      FreshRSS_Context::$user_conf->oai_temperature = (float)Minz_Request::param('oai_temperature', 0.7);
       FreshRSS_Context::$user_conf->save();
     }
   }
